Add names; Class, race values come from input txt
The program will now give the character a name when the
program is run - names come from input txt file "names.txt"

I have done a small refactor so that the race and class values
are also now read from input files - race.txt and class.txt.

I've done this so that the user can add values at will without
having to futz about with the program internals. This is happy-making!