Fleshed out variable names. Fixed bug.
Fixed a bug where I was not setting the character race or background a
single time for the entire program run. As a result, each method call
was using a different rand under the covers, which we don't want
because we want race in each run to be static. Same for background.